Created test zotero .bib.test file
The name of the file is EROI_review_v2_zotero.bib.test. I added the
.test on the end so that it would confuse Tex. Not sure if this is going
to work because  it creates its own labels upon export. Helpfully, it is
ALWAYS the same format: "author last name_first word of title_year." If
that is not unique, it will add a -1, -2, etc. to the next citation(s)
codename. For example,
@article{georgescu-roegen_energy_1979,
 title = {Energy and matter in
mankind's technological circuit},
 volume = {10},
 journal = {Journal of
Business Administration},
 author = {Georgescu-Roegen, N.},
 year =
{1979},
 pages = {107–127}
},

@article{georgescu-roegen_energy_1979-1,

title = {Energy analysis and economic valuation},
 journal = {Southern
Economic Journal},
 author = {Georgescu-Roegen, N.},
 year = {1979},

pages = {1023–1058}
},